+++ src/card/cards/combat_constraints.rs
@@ -93,3 +93,7 @@ pub const SOUL_OF_THE_HARVEST: CardDefinitionId = CardDefinitionId(1892);
 pub const RAT_TOKEN_1_1_BLACK: CardDefinitionId = CardDefinitionId(1893);
 pub const OGRE_SLUMLORD: CardDefinitionId = CardDefinitionId(1894);
 pub const SUBLIME_ARCHANGEL: CardDefinitionId = CardDefinitionId(1895);
+pub const PONTIFF_OF_BLIGHT: CardDefinitionId = CardDefinitionId(1896);
+pub const BATTERING_KRASIS: CardDefinitionId = CardDefinitionId(1897);
+pub const EMMARA_TANDRIS: CardDefinitionId = CardDefinitionId(1898);
+pub const TITHE_DRINKER: CardDefinitionId = CardDefinitionId(1899);